feat: Add Ansible playbook and inventory for Kubernetes cluster configuration
This commit introduces a new Ansible playbook to configure static IP addresses for the Kubernetes master and worker nodes, as well as update the /etc/hosts file with the cluster node information. Additionally, an inventory file is created to define the cluster's host details and SSH configuration, facilitating automated deployment and management of the Kubernetes environment.